+++ src/main/java/org/truffleruby/parser/ast/OpElementAsgnParseNode.java
@@ -44,7 +44,7 @@ import org.truffleruby.parser.ast.visitor.NodeVisitor;
  * a[4] += 5
  * a[3] &&= true
  * </pre>
-*/
+ */
 public class OpElementAsgnParseNode extends ParseNode {
     private final ParseNode receiverNode;
     private final ParseNode argsNode;