+++ CHANGELOG.md
@@ -58,6 +58,7 @@ Performance:
 
 Changes:
 
+* TruffleRuby now requires Java 11+ and no longer supports Java 8.
 
 # 21.3.0